Benchmarking/requirements.txt
google-generative-ai>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.1.0
numpy>=1.18.0